pyproject.toml
[project]
name = "personachat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.18.0",
    "python-dotenv>=1.1.0",
]
